fix(download): 修复编译错误——anyFailure 作用域提到 try 外 + baseBps 显式转 long
- CNHotUpdateCheck: anyFailure 在 try 内声明但末尾 if 在 try 外 → 提到方法级
- CNChunkedDownload: baseBytes/(double) 赋 long 缺 (long) 转换
- 本地 javac(android-all.jar + RestClient stub)验证通过

+++ patch/src/main/java/io/kamihama/magianative/CNChunkedDownload.java
@@ -412,7 +412,7 @@ public final class CNChunkedDownload {
                         }
                         if (sinceStart > baseToNs) {
                             if (baseTimeNs > 0) {
-                                baseBps = baseBytes / (baseTimeNs / 1_000_000_000.0);
+                                baseBps = (long)(baseBytes / (baseTimeNs / 1_000_000_000.0));
                             }
                             if (baseBps > 0) CNMirrors.reportBaseline(mirror, baseBps);
                             baseReady = true;
